+++ src/test/java/com/example/demo/utils/DatabaseCleanup.java
@@ -1,5 +1,9 @@
 package com.example.demo.utils;
 
+import static org.assertj.core.api.Assertions.*;
+
+import java.time.LocalDateTime;
+import java.time.LocalTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -9,8 +13,12 @@ import javax.persistence.PersistenceContext;
 import javax.persistence.Table;
 import javax.persistence.metamodel.EntityType;
 
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import com.example.demo.util.TruncatedTimeUtil;
 
 @Service
 public class DatabaseCleanup {
@@ -51,5 +59,22 @@ public class DatabaseCleanup {
 			return null;
 		}
 		return input.replaceAll("([a-z])([A-Z]+)", "$1_$2").toLowerCase();
+	}
+
+	static class TruncatedTimeUtilTest {
+
+		@Test
+		@DisplayName("나노초에서 초단위로 절삭한다.")
+		void convert_nano_to_micro_time() {
+			LocalTime result = TruncatedTimeUtil.truncateTimeToSecond(LocalTime.of(23, 59, 59, 999_999_999));
+			assertThat(result).isEqualTo(LocalTime.of(23, 59, 59));
+		}
+
+		@Test
+		@DisplayName("나노초에서 초단위로 절삭한다.")
+		void convert_nano_to_micro_date_time() {
+			LocalDateTime result = TruncatedTimeUtil.truncateDateTimeToSecond(LocalDateTime.MAX);
+			assertThat(result).isEqualTo(LocalDateTime.of(999999999, 12, 31, 23, 59, 59));
+		}
 	}
 }
